// src/features/personEditForm/types/formFieldNames.ts
export enum FormFieldNames {
	Name = 'name',
	BirthYear = 'birth_year',
	EyeColor = 'eye_color',
	Gender = 'gender',
	HairColor = 'hair_color',
	Height = 'height',
	Mass = 'mass',
	SkinColor = 'skin_color',
	Homeworld = 'homeworld',
	Films = 'films',
	Species = 'species',
	Starships = 'starships',
	Vehicles = 'vehicles',
	Url = 'url',
	Created = 'created',
	Edited = 'edited',
}